pyproject.toml
[project]
name = "project-3-astrarag-agentic-rag-chatbot-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb==1.0.21",
    "crewai==0.186.1",
    "fastapi==0.116.1",
    "llama-index>=0.14.3",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-groq>=0.4.1",
    "llama-index-vector-stores-chroma>=0.5.5",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "streamlit==1.49.1",
    "uvicorn==0.35.0",
]
